逗猫页音频播放功能

+++ CatTranslator/Feature/GetWidthHeight.swift
@@ -16,11 +16,11 @@ class G {
     let IPhone11ProWidth = 375.0
     let IPhone11ProHeigth = 812.0
     
-    func w(_ width:Float) -> Float{
-        return Float(fullScreenSize.width) * (width / Float(IPhone11ProWidth))
+    func w(_ width:Float) -> CGFloat{
+        return CGFloat(Float(fullScreenSize.width) * width / Float(IPhone11ProWidth))
     }
     
-    func h(_ height:Float) -> Float {
-        return Float(fullScreenSize.height) * ( height / Float(IPhone11ProHeigth))
+    func h(_ height:Float) -> CGFloat {
+        return CGFloat(Float(fullScreenSize.height) * ( height / Float(IPhone11ProHeigth)))
     }
 }
